Extract ExecutionResult and StepResult from src-cli
